requirements.txt
asn1crypto==0.24.0
cachetools==3.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.5
ecdsa==0.13
Flask==1.0.2
Flask-Login==0.4.1
future==0.17.1
google-api-core==1.7.0
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-datastore==1.7.1
googleapis-common-protos==1.5.6
grpcio==1.18.0
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jwcrypto==0.6.0
MarkupSafe==1.1.0
Naked==0.1.31
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycrypto==2.6.1
python-dateutil==2.6.1
python-jose==3.0.1
python-jwt==3.2.4
pytz==2018.9
PyYAML==3.13
requests==2.20.1
rsa==4.0
shellescape==3.4.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
